Move common cache types to a separate file
